requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pytube==15.0.0
requests==2.31.0
youtube-transcript-api==0.6.2
python-multipart==0.0.9 